Add files via upload
The goal of this assignment is to test your knowledge of: HTML, CSS (through Sass) and JavaScript.
You need to build a fully responsive html page following the designs & guide below. To make things simpler, we have
split the guide into 3 main sections: Hero, Features & Pricing. We will cover the requirements for each section
separately. All the assets will be available in a zip file attached to the email with the assignment.